Track paths in messages.
Each message carries with it, for accounting purposes, all the paths it
took to reach the current node. This is a problem because the number of
paths tends to get very large and simulations are slow. We will save
this here and then go on to represent paths in a graph.